// frontend/src/api/generated/models/partner-detail.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { PartnerType } from './partner-type';

/**
 * 
 * @export
 * @interface PartnerDetail
 */
export interface PartnerDetail {
    /**
     * 
     * @type {number}
     * @memberof PartnerDetail
     */
    'id': number;
    /**
     * 
     * @type {string}
     * @memberof PartnerDetail
     */
    'partnerCode': string;
    /**
     * 
     * @type {string}
     * @memberof PartnerDetail
     */
    'partnerName': string;
    /**
     * 
     * @type {string}
     * @memberof PartnerDetail
     */
    'partnerNameKana'?: string;
    /**
     * 
     * @type {PartnerType}
     * @memberof PartnerDetail
     */
    'partnerType': PartnerType;
    /**
     * 
     * @type {string}
     * @memberof PartnerDetail
     */
    'address'?: string;
    /**
     * 
     * @type {string}
     * @memberof PartnerDetail
     */
    'phone'?: string;
    /**
     * 
     * @type {string}
     * @memberof PartnerDetail
     */
    'contactPerson'?: string;
    /**
     * 
     * @type {string}
     * @memberof PartnerDetail
     */
    'email'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof PartnerDetail
     */
    'isActive': boolean;
    /**
     * 
     * @type {number}
     * @memberof PartnerDetail
     */
    'version': number;
    /**
     * 
     * @type {string}
     * @memberof PartnerDetail
     */
    'createdAt': string;
    /**
     * 
     * @type {string}
     * @memberof PartnerDetail
     */
    'updatedAt': string;
}