// frontend/src/api/generated/models/partner-type.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 取引先種別（仕入先・出荷先・両方）
 * @export
 * @enum {string}
 */

export const PartnerType = {
    Supplier: 'SUPPLIER',
    Customer: 'CUSTOMER',
    Both: 'BOTH'
} as const;

export type PartnerType = typeof PartnerType[keyof typeof PartnerType];